+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added compatibility with provisional drafts in Craft 3.7
+- Added the ability to copy and paste block types between fields
 
 ### Changed
 - Neo now requires Craft 3.7.0 beta 3 or later